Update to latest open ai, langchain and openai
