+++ packages/aila/package.json
@@ -16,6 +16,7 @@
   },
   "prettier": "@oakai/prettier-config",
   "dependencies": {
+    "@ai-sdk/openai": "^0.0.55",
     "@clerk/nextjs": "5.1.0",
     "@oakai/core": "*",
     "@oakai/db": "*",
@@ -23,11 +24,13 @@
     "@oakai/logger": "*",
     "@sentry/nextjs": "^8.19.0",
     "@vercel/kv": "^0.2.2",
+    "ai": "^3.3.26",
     "american-british-english-translator": "^0.2.1",
     "cloudinary": "^1.41.1",
     "dotenv-cli": "^6.0.0",
     "jsonrepair": "^3.8.0",
-    "openai": "^4.52.0",
+    "langchain": "^0.2.17",
+    "openai": "^4.57.1",
     "remeda": "^1.29.0",
     "superjson": "^1.9.1",
     "tiny-invariant": "^1.3.1",
